+++ pkg/filters/filters.go
@@ -1,0 +1,17 @@
+package filters
+
+import "math"
+
+const (
+	filterNotEqual uint32 = iota
+	filterEqual
+)
+
+// Set default inequality values
+// val<0 and val>math.MaxUint64 should never be used by the user as they give an empty set
+const (
+	LessNotSetUint    uint64 = 0
+	GreaterNotSetUint uint64 = math.MaxUint64
+	LessNotSetInt     int64  = math.MinInt64
+	GreaterNotSetInt  int64  = math.MaxInt64
+)
